# docs/openapi/spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Financeira",
    "version": "1.0.0",
    "description": "API de gestão financeiros, orientada a eventos."
  },
  "servers": [
    {
      "url": "http://localhost:8082",
      "description": "Local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/finance/expense/{competenceDate}": {
      "get": {
        "tags": ["Expense"],
        "summary": "Listar despesas paginadas",
        "operationId": "getExpense",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/CompetenceDate" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Size" }
        ],
        "responses": {
          "200": {
            "description": "Listar despesas paginadas",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExpensePageResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "post": {
        "tags": ["Expense"],
        "summary": "Adicionar despesa",
        "operationId": "addExpense",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/CompetenceDate" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Size" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddExpenseRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Despesa adicionada. Retorna lista paginada atualizada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExpensePageResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "put": {
        "tags": ["Expense"],
        "summary": "Editar detalhe salarial",
        "operationId": "updateExpense",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/CompetenceDate" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Size" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateExpenseRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Despesa atualizada. Retorna lista paginada atualizada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExpensePageResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      },
      "delete": {
        "tags": ["Expense"],
        "summary": "Deletar um ou mais despesas",
        "operationId": "deleteExpense",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/CompetenceDate" },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Size" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DeleteExpenseRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Despesas deletadas. Retorna lista paginada atualizada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExpensePageResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido via Keycloak."
      }
    },
    "parameters": {
      "CompetenceDate": {
        "name": "competenceDate",
        "in": "path",
        "required": true,
        "description": "Competência no formato yyyy-MM-dd",
        "schema": {
          "type": "string",
          "format": "date",
          "example": "2026-06-01"
        }
      },
      "Page": {
        "name": "page",
        "in": "query",
        "required": false,
        "description": "Número da página (0-based)",
        "schema": { "type": "integer", "default": 0, "example": 0 }
      },
      "Size": {
        "name": "size",
        "in": "query",
        "required": false,
        "description": "Quantidade de itens por página",
        "schema": { "type": "integer", "default": 10, "example": 10 }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Requisição inválida",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorViolacaoResponse"
            },
            "example": {
              "status": 400,
              "error": "Requisição inválida",
              "message": "Salário líquido não pode ser maior que o salário bruto",
              "timestamp": "2026-05-03T19:00:00.000Z",
              "path": "/api/v1/finance/summary/2026-06-01",
              "violacoes": [
                {
                  "campo": "userNetSalary",
                  "razao": "Salário líquido não pode ser maior que o salário bruto"
                },
                {
                  "campo": "userGrossSalary",
                  "razao": "Salário bruto não pode ser menor que o salário líquido"
                }
              ]
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Falha na autenticação. Token ausente, inválido ou expirado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 401,
              "error": "Não autorizado",
              "message": "Falha na autenticação. Token ausente, inválido ou expirado",
              "timestamp": "2026-05-03T19:00:00.000Z",
              "path": "/api/v1/ficance"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Acesso negado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 403,
              "error": "Acesso negado",
              "message": "Você não tem permissão para acessar este recurso",
              "timestamp": "2026-05-03T19:00:00.000Z",
              "path": "/api/v1/ficance"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Erro interno do servidor",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "timestamp": "2026-05-03T19:00:00.000Z",
              "status": 500,
              "error": "Erro interno do servidor",
              "message": "Ocorreu um erro inesperado. Por favor, tente novamente mais tarde",
              "path": "/api/v1/ficance"
            }
          }
        }
      },
      "NotFound": {
        "description": "Não Encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "status": 404,
              "error": "Não Encontrado",
              "message": "Objeto não encontrado",
              "timestamp": "2026-05-03T19:00:00.000Z",
              "path": "/api/v1/ficance"
            }
          }
        }
      }
    },
    "schemas": {
...
      "CardResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64", "description": "ID do cartão", "example": 1 },
          "name": { "type": "string", "description": "Nome do cartão", "example": "Cartão Visa" },
          "debit": { "type": "boolean", "description": "Indica se o cartão é de débito ou crédito (true=Débito, false=Crédito)", "example": false },
          "number": { "type": "string", "description": "Número do cartao", "example": "241523436122" },
          "due": { "type": "string", "format": "date", "description": "Data de vencimento", "example": "2027-02-10" }
        }
      },
      "PaymentResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64", "description": "ID do pagamento", "example": 1 },
          "method": { "type": "integer", "format": "int64", "description": "Indica o metodo de pagamento  (1=Cartão, 2=Pix, 3=Boleto, 4=Dinheiro)", "example": 1 },
          "paid": { "type": "string", "format": "date", "description": "Data do pagamento", "example": "2026-05-10" },
          "detail": { "type": "string", "description": "Detalhe do pagamento", "example": "Parcela 3/5" },
          "card": {
            "$ref": "#/components/schemas/CardResponse"
          }
        }
      },
      "ExpenseResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64", "description": "ID da despesa", "example": 1 },
          "due": { "type": "string", "format": "date", "description": "Data de vencimento", "example": "2026-05-03" },
          "shared": { "type": "boolean", "description": "Indica se a despesa é compartilhada (true=Familiar, false=Individual)", "example": false },
          "name": { "type": "string", "description": "Nome da despesa", "example": "Energia" },
          "amount": { "type": "number", "format": "decimal", "description": "Valor da despesa", "example": 540.35 },
          "type": { "type": "integer", "format": "int64","description": "ID da despesa (1=Custo Fixo, 2=Conforto, 3=Metas, 4=Prazeres, 5=Investimento, 6=Conhecimento)", "example": 1 },
          "detail": { "type": "string", "description": "Detalhe da despesa", "example": "Consumo de energia do apartamento" },
          "payment": {
            "$ref": "#/components/schemas/PaymentResponse"
          }
        }
      },
      "ExpensePageResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ExpenseResponse" }
          },
          "totalElements": { "type": "integer", "format": "int64", "example": 3 },
          "totalPages": { "type": "integer", "example": 1 },
          "size": { "type": "integer", "example": 10 },
          "number": { "type": "integer", "example": 0 }
        }
      },
      "AddCardRequest": {
        "type": "object",
        "required": ["name", "debit", "number", "due"],
        "properties": {
          "id": { "type": "integer", "format": "int64", "description": "ID do cartão", "example": 1 },
          "name": { "type": "string", "description": "Nome do cartão", "example": "Cartão Visa" },
          "debit": { "type": "boolean", "description": "Indica se o cartão é de débito ou crédito (true=Débito, false=Crédito)", "example": false },
          "number": { "type": "string", "description": "Número do cartao", "example": "241523436122" },
          "due": { "type": "string", "format": "date", "description": "Data de vencimento", "example": "2027-02-10" }
        }
      },
      "AddPaymentRequest": {
        "type": "object",
        "required": ["method", "paid"],
        "properties": {
          "id": { "type": "integer", "format": "int64", "description": "ID do pagamento", "example": 1 },
          "method": { "type": "integer", "format": "int64", "description": "Indica o metodo de pagamento  (1=Cartão, 2=Pix, 3=Boleto, 4=Dinheiro)", "example": 1 },
          "paid": { "type": "string", "format": "date", "description": "Data do pagamento", "example": "2026-05-10" },
          "detail": { "type": "string", "description": "Detalhe do pagamento", "example": "Parcela 3/5" },
          "card": {
            "$ref": "#/components/schemas/AddCardRequest"
          }
        }
      },
      "AddExpenseRequest": {
        "type": "object",
        "required": ["due", "shared", "name", "amount", "type"],
        "properties": {
          "due": { "type": "string", "format": "date", "description": "Data de vencimento", "example": "2026-05-03" },
          "shared": { "type": "boolean", "description": "Indica se a despesa é compartilhada (true=Familiar, false=Individual)", "example": false },
          "name": { "type": "string", "description": "Nome da despesa", "example": "Energia" },
          "amount": { "type": "number", "format": "decimal", "description": "Valor da despesa", "example": 540.35 },
          "type": { "type": "integer", "format": "int64","description": "ID da despesa (1=Custo Fixo, 2=Conforto, 3=Metas, 4=Prazeres, 5=Investimento, 6=Conhecimento)", "example": 1 },
          "detail": { "type": "string", "description": "Detalhe da despesa", "example": "Consumo de energia do apartamento" },
          "payment": {
            "$ref": "#/components/schemas/AddPaymentRequest"
          }
        }
      },
      "UpdateExpenseRequest": {
        "type": "object",
        "required": ["id", "due", "shared", "name", "amount", "type"],
        "properties": {
          "id": { "type": "integer", "format": "int64", "description": "ID da despesa", "example": 1 },
          "due": { "type": "string", "format": "date", "description": "Data de vencimento", "example": "2026-05-03" },
          "shared": { "type": "boolean", "description": "Indica se a despesa é compartilhada (true=Familiar, false=Individual)", "example": false },
          "name": { "type": "string", "description": "Nome da despesa", "example": "Energia" },
          "amount": { "type": "number", "format": "decimal", "description": "Valor da despesa", "example": 540.35 },
          "type": { "type": "integer", "format": "int64","description": "ID da despesa (1=Custo Fixo, 2=Conforto, 3=Metas, 4=Prazeres, 5=Investimento, 6=Conhecimento)", "example": 1 },
          "detail": { "type": "string", "description": "Detalhe da despesa", "example": "Consumo de energia do apartamento" },
          "payment": {
            "$ref": "#/components/schemas/AddPaymentRequest"
          }
        }
      },
      "DeleteExpenseRequest": {
        "type": "object",
        "required": ["ids"],
        "properties": {
          "ids": {
            "type": "array",
            "minItems": 1,
            "items": { "type": "integer", "format": "int64" },
            "description": "Lista de IDs de expense a deletar",
            "example": [1, 2]
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Data e hora do erro em UTC",
            "example": "2026-05-03T19:00:00.000Z"
          },
          "status": {
            "type": "integer",
            "description": "Código HTTP do erro",
            "example": 400
          },
          "error": {
            "type": "string",
            "description": "Descrição resumida do erro HTTP",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "description": "Mensagem detalhada do erro",
            "example": "Mensagem de erro"
          },
          "path": {
            "type": "string",
            "description": "Caminho do endpoint que gerou o erro",
            "example": "/api/v1/finance"
          }
        }
      },
      "ErrorViolacaoResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Data e hora do erro em UTC"
          },
          "status": {
            "type": "integer",
            "description": "Código HTTP do erro",
            "example": 400
          },
          "error": {
            "type": "string",
            "description": "Descrição resumida do erro HTTP",
            "example": "Bad Request"
          },
          "violacoes": {
            "type": "array",
            "description": "Lista de violações de regras de negócio que causaram o erro",
            "items": {
              "$ref": "#/components/schemas/ViolacaoResponse"
            }
          },
          "path": {
            "type": "string",
            "description": "Caminho do endpoint que gerou o erro",
            "example": "/api/v1/ficance"
          },
          "message": {
            "type": "string",
            "description": "Mensagem detalhada do erro",
            "example": "Mensagem de erro"
          }
        }
      },
      "ViolacaoResponse": {
        "type": "object",
        "properties": {
          "campo": {
            "type": "string",
            "description": "Campo que violou a regra de negócio"
          },
          "razao": {
            "type": "string",
            "description": "Razão da violação de regra de negócio"
          }
        }
      }
    }
  }
}